backend/requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-Migrate>=4.0,<5
marshmallow>=3.20,<4
marshmallow-sqlalchemy>=1.0,<2
python-dotenv>=1.0,<2
gunicorn>=22.0,<24
